frontend: split out the image presentation functionality
Also changes the icon that appears for all images to be the same
as the priority images (generic icon for images)
